Moved Header compnent to its own module, imported into app
